refactor(apps): cross-app Tier 3 — shared overflow hook + app-component stub removal
- Extract useTextOverflow hook; adopt in LcdAnimatedTitle and TV ScrollingChannelName (output-identical marquee overflow dedup)
- Remove 11 app-component re-export stubs (Videos/ControlPanels/Contacts/Finder/Tv/Calendar/Admin/Chats/Synth/Maps/InternetExplorer); repoint appRegistry lazy imports + app index files to the real *-app/ module paths
