--- src/apps/admin/components/AdminAppComponent.tsx
@@ -1,1 +1,0 @@
-export { AdminAppComponent } from "./admin-app/AdminAppComponent";
