--- src/apps/calendar/components/CalendarAppComponent.tsx
@@ -1,1 +1,0 @@
-export { CalendarAppComponent } from "./calendar-app/CalendarAppComponent";
