--- src/apps/chats/components/ChatsAppComponent.tsx
@@ -1,1 +1,0 @@
-export { ChatsAppComponent } from "./chats-app/ChatsAppComponent";
