--- src/apps/contacts/components/ContactsAppComponent.tsx
@@ -1,1 +1,0 @@
-export { ContactsAppComponent } from "./contacts-app/ContactsAppComponent";
