--- src/apps/control-panels/components/ControlPanelsAppComponent.tsx
@@ -1,1 +1,0 @@
-export { ControlPanelsAppComponent } from "./control-panels-app/ControlPanelsAppComponent";
