--- src/apps/finder/components/FinderAppComponent.tsx
@@ -1,1 +1,0 @@
-export { FinderAppComponent } from "./finder-app/FinderAppComponent";
