--- src/apps/internet-explorer/components/InternetExplorerAppComponent.tsx
@@ -1,1 +1,0 @@
-export { InternetExplorerAppComponent } from "./internet-explorer-app/InternetExplorerAppComponent";
