--- src/apps/maps/components/MapsAppComponent.tsx
@@ -1,1 +1,0 @@
-export { MapsAppComponent } from "./maps-app/MapsAppComponent";
