--- src/apps/synth/components/SynthAppComponent.tsx
@@ -1,1 +1,0 @@
-export { SynthAppComponent } from "./synth-app/SynthAppComponent";
