--- src/apps/tv/components/TvAppComponent.tsx
@@ -1,1 +1,0 @@
-export { TvAppComponent } from "./tv-app/TvAppComponent";
