--- src/apps/videos/components/VideosAppComponent.tsx
@@ -1,1 +1,0 @@
-export { VideosAppComponent } from "./videos-app/VideosAppComponent";
